refactor: retryUntilValid view로 이동
